requirements.txt
# requirements.txt
sentence-transformers
faiss-cpu
ollama
striprtf
tqdm
numpy
torch
sqlite3
requests